+++ docs/source/changelog.rst
@@ -11,6 +11,7 @@ Upcoming Release
 
 - Fix bug when ``load()`` is called without any arguments (:pr:`3`)
 - Internal refactoring to improve test coverage (:pr:`3`)
+- Update to use ``requests.get`` URL parameters (:pr:`4`)
 
 Version 0.1.0 (November 30, 2018)
 ---------------------------------